cleanup header includes and orders
Not complete, but a good start.
