+++ arm7/lib/include/SND_exChannel.h
@@ -1,0 +1,8 @@
+#ifndef GUARD_SND_EXCHANNEL_H
+#define GUARD_SND_EXCHANNEL_H
+
+#include "global.h"
+
+bool SND_IsChannelActive(u32 idx);
+
+#endif //GUARD_SND_EXCHANNEL_H
