+++ arm9/lib/include/systemWork.h
@@ -1,7 +1,6 @@
 #ifndef POKEDIAMOND_ARM9_SYSTEMWORK_H
 #define POKEDIAMOND_ARM9_SYSTEMWORK_H
 
-#include "OS_context.h"
 #include "nitro/OS_systemWork_shared.h"
 
 #endif //POKEDIAMOND_ARM9_SYSTEMWORK_H
