+++ arm9/lib/src/MI_wram.c
@@ -1,6 +1,7 @@
-#include "function_target.h"
-#include "consts.h"
 #include "MI_wram.h"
+
+#include "global.h"
+#include "registers.h"
 
 ARM_FUNC void MI_SetWramBank(MIWram cnt)
 {
